Atualização - Aula 5 - int para Double

+++ Aula_05/AvaliarAlunos.java
@@ -5,10 +5,10 @@ public class AvaliarAlunos {
         System.out.print("Insira o número de estudantes: ");
         int numEstudantes = entrada.nextInt();
 
-        int[] vetor = new int[numEstudantes];
+        double[] vetor = new double[numEstudantes];
         for (int i=0; i < vetor.length; i++){
             System.out.printf("Estudante %2d: ", i + 1);
-            vetor[i] = entrada.nextInt();
+            vetor[i] = entrada.nextDouble();
         }        
         double media = calcularMediaTurma(vetor);
         System.out.println("Média das notas: " + media);
@@ -16,7 +16,7 @@ public class AvaliarAlunos {
 
     }
 
-    public static double calcularMediaTurma(int[] vetor){
+    public static double calcularMediaTurma(double[] vetor){
        double soma = 0;
        int qtde = 0;
        for (int i = 0; i < vetor.length; i++){
@@ -26,7 +26,7 @@ public class AvaliarAlunos {
        return soma/qtde;
     }
 
-    public static void imprimirResultados(int[] vetor){
+    public static void imprimirResultados(double[] vetor){
        for (int i = 0; i < vetor.length; i++){
             if (vetor[i] < 7){
                 System.out.printf("Estudante %2d está de recuperação!\n", i + 1);
